feat :  홈 페이지 병합, 수정
